Move Polish bank levy asset perimeter out of Banking
